1. modify - Repository - Spectator
