+++ app/src/main/java/com/plz/no/anr/lol_usersearch_compose/ui/feature/spectator/composables/SpectatorView.kt
@@ -1,0 +1,2 @@
+package com.plz.no.anr.lol_usersearch_compose.ui.feature.spectator.composables
+
